+++ vector/src/main/java/im/vector/riotx/features/roommemberprofile/RoomMemberProfileViewEvents.kt
@@ -31,4 +31,6 @@ sealed class RoomMemberProfileViewEvents : VectorViewEvents {
             val userId: String,
             val canCrossSign: Boolean
     ) : RoomMemberProfileViewEvents()
+
+    object ShareRoomMemberProfile : RoomMemberProfileViewEvents()
 }